+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 3.0.0
 
 - Change too-tall detection to exceeds-viewport.
+- Make exceeds viewport detection re-compute on window resize
 
 ## 2.3.3